src/main/java/no/nav/brukernotifikasjon/schemas/builders/DoneBuilder.java
package no.nav.brukernotifikasjon.schemas.builders;


import no.nav.brukernotifikasjon.schemas.Done;
import no.nav.brukernotifikasjon.schemas.builders.util.ValidationUtil;

import java.time.LocalDateTime;

public class DoneBuilder {

    private LocalDateTime tidspunkt;

    public DoneBuilder withTidspunkt(LocalDateTime tidspunkt) {
        this.tidspunkt = tidspunkt;
        return this;
    }

    public Done build() {
        return new Done(
                ValidationUtil.localDateTimeToUtcTimestamp(tidspunkt, "tidspunkt", ValidationUtil.IS_REQUIRED_TIDSPUNKT)
        );
    }
}
